+++ src/pages/myPlayList/index.tsx
@@ -1,83 +1,28 @@
-import { useAuthStore } from '../../store/useAuthStore'
-import { supabase } from '@/supabase/supabaseConfig'
 import { useToast } from '@/hooks/useToast'
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
-import { useNavigate } from 'react-router-dom'
-
 import { PlayList, Loading } from '@/component'
 import * as S from '@/component/MyPlayList/MyPlayList.styles'
 import { IPlayListType } from '@/types/playList'
-
-const fetchMyPlaylists = async (userId: string): Promise<IPlayListType[]> => {
-  const { data, error } = await supabase.rpc('get_my_playlists', {
-    p_user_id: userId
-  })
-  if (error) throw new Error('나의 플레이리스트를 불러오는데 실패했습니다')
-  return data || []
-}
-
-const deletePlaylist = async (playlistId: string) => {
-  const { error } = await supabase.rpc('delete_playlist', {
-    p_playlist_id: playlistId
-  })
-  if (error) throw new Error('플레이리스트 삭제에 실패했습니다.')
-}
+import { useMyPlayList } from '@/hooks/useMyPlayList'
 
 export const MyPlayLists = () => {
-  const user = useAuthStore(state => state.user)
-  const { handleToastError, handleToastSuccess } = useToast()
-  const queryClient = useQueryClient()
-  const navigate = useNavigate()
-
   const {
-    data: playlists = [],
+    playlists,
     isLoading,
-    error
-  } = useQuery({
-    queryKey: ['myPlaylists', user?.id],
-    queryFn: async () => {
-      if (!user?.id) return []
-      return await fetchMyPlaylists(String(user?.id))
-    },
-    enabled: !!user?.id,
-    retry: 2
-  })
+    error,
+    handleDeletePlayList,
+    handleDetailPage,
+    user
+  } = useMyPlayList()
+  const { handleToastError } = useToast()
+
+  if (isLoading) {
+    return <Loading />
+  }
 
   if (error) {
     handleToastError(
       (error as Error).message || '나의 플레이리스트를 불러오는데 실패했습니다.'
     )
-  }
-
-  const deleteMutation = useMutation({
-    mutationFn: async (deletedId: string) => {
-      await deletePlaylist(deletedId)
-      return deletedId
-    },
-    onSuccess: deletedId => {
-      queryClient.setQueryData(
-        ['myPlaylists', user?.id],
-        (old: IPlayListType[] | undefined) =>
-          old?.filter(playlist => playlist.playlist_id !== deletedId)
-      )
-      queryClient.invalidateQueries({ queryKey: ['playList'] })
-      handleToastSuccess('플레이리스트가 삭제되었습니다.')
-      navigate('/', { state: { refetch: true } })
-    },
-    onError: () => {
-      handleToastError('플레이리스트 삭제에 실패했습니다.')
-    }
-  })
-
-  const handleDeletePlayList = (deletedId: string) => {
-    deleteMutation.mutate(deletedId)
-  }
-
-  if (isLoading) {
-    return <Loading />
-  }
-  const handleDetailPage = (playlistId: string) => {
-    navigate(`/view/${playlistId}`)
   }
 
   return (
